# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tadweerah API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listMarketplaceListings
      tags: [listings]
      summary: Public marketplace listings (open status only). Buyers browse here.
      parameters:
        - in: query
          name: material
          schema:
            $ref: "#/components/schemas/WasteMaterial"
        - in: query
          name: city
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteListing"
        "401":
          description: Unauthorized
        "403":
          description: Buyer access required
    post:
      operationId: createWasteListing
      tags: [listings]
      summary: Create a waste listing for the current producer company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWasteListingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteListing"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Producer access required
  /listings/mine:
    get:
      operationId: listMyListings
      tags: [listings]
      summary: Listings owned by the current producer company
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/WasteListingStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteListing"
        "401":
          description: Unauthorized
        "403":
          description: Producer access required
  /listings/{waste_listing_id}:
    get:
      operationId: getWasteListing
      tags: [listings]
      summary: Get a single listing by id
      parameters:
        - in: path
          name: waste_listing_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteListing"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /listings/{waste_listing_id}/close:
    post:
      operationId: closeWasteListing
      tags: [listings]
      summary: Close a listing (owner producer only)
      parameters:
        - in: path
          name: waste_listing_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteListing"
        "401":
          description: Unauthorized
        "403":
          description: Not the owner
        "404":
          description: Not found
...
components:
  schemas:
...
    SaleType:
      type: string
      enum: [auction, direct]
      description: |
        auction = open bidding, all eligible buyers can submit offers.
        direct = targeted sale; seller controls who can see and participate.
    WasteMaterial:
      type: string
      enum: [paper, plastic, metal, glass, electronics, organic, other]
    WasteUnit:
      type: string
      enum: [kg, ton]
    WasteListingStatus:
      type: string
      enum: [open, closed]
    ListingVisibility:
      type: string
      enum: [public, private]
      description: |
        Controls whether a listing appears in the public marketplace feed.
        CURRENTLY READ-ONLY: all listings default to "public".
    PricingModel:
      type: string
      enum: [fixed, by_weight, revenue_share]
      description: |
        Governs how price_per_unit is interpreted and what settlement mechanics apply.
        Immutable once a listing is published.
        fixed: price_per_unit × quantity = agreed commercial amount.
        by_weight: price_per_unit is a rate; final amount settled post-weighing.
        revenue_share: buyer pays a percentage of their resale revenue (direct sale only).
          revenue_share_pct field required when this model is selected.
    WasteListing:
      type: object
      properties:
        id:
          type: string
        company_id:
          type: string
        company_name:
          type: string
        material:
          $ref: "#/components/schemas/WasteMaterial"
        quantity:
          type: number
        unit:
          $ref: "#/components/schemas/WasteUnit"
        city:
          type: string
        description:
          type: string
        price_hint:
          type: number
        status:
          $ref: "#/components/schemas/WasteListingStatus"
        pricing_model:
          $ref: "#/components/schemas/PricingModel"
        sale_type:
          $ref: "#/components/schemas/SaleType"
        material_category_id:
          type: string
          description: FK to material_categories.id. Null for legacy listings.
        material_subcategory_id:
          type: string
          description: FK to material_categories.id (sub-level). Null if not specified.
        unit_option_id:
          type: string
          description: FK to unit_options.id. Null for listings using legacy unit enum.
        revenue_share_pct:
          type: number
          description: |
            Revenue sharing percentage (0–100). Only present when pricing_model = revenue_share.
            Represents the percentage of the buyer's resale revenue paid to the producer.
        required_services:
          type: array
          description: >
            Capabilities a buyer must have to submit an offer. Empty array means no restrictions.
          items:
            $ref: "#/components/schemas/ListingRequiredService"
        visibility:
          $ref: "#/components/schemas/ListingVisibility"
        my_offer_price:
          type: number
          description: >
            The calling buyer's current offer price_per_unit on this listing.
            Only included in GET /listings (marketplace) responses.
            Null/absent if the buyer has no offer.
        my_rank:
          type: integer
          description: >
            The calling buyer's current rank among all offers (1 = highest).
            Only included in GET /listings (marketplace) responses.
            Null/absent if the buyer has no offer.
        created_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        offer_count:
          type: integer
          description: Total number of active (non-withdrawn) offers on this listing.
        highest_offer_total:
          type: number
          description: Highest offer as a total (price_per_unit × quantity). Null if no offers.
      required:
        [id, company_id, company_name, material, quantity, unit, city, status, created_at]
    CreateWasteListingBody:
      type: object
      properties:
        material:
          $ref: "#/components/schemas/WasteMaterial"
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          $ref: "#/components/schemas/WasteUnit"
        city:
          type: string
          minLength: 2
          maxLength: 80
        description:
          type: string
          maxLength: 500
        price_hint:
          type: number
          minimum: 0
        pricing_model:
          $ref: "#/components/schemas/PricingModel"
          description: |
            Governs settlement mechanics. Immutable once published.
            Defaults to "fixed" if not supplied.
        sale_type:
          $ref: "#/components/schemas/SaleType"
          description: |
            Controls offer model. Defaults to "auction" if not supplied.
        material_category_id:
          type: string
          description: Optional FK to material_categories.id.
        material_subcategory_id:
          type: string
          description: Optional FK to material_categories.id (sub-level).
        unit_option_id:
          type: string
          description: Optional FK to unit_options.id. Falls back to unit enum if absent.
        revenue_share_pct:
          type: number
          minimum: 0
          maximum: 100
          description: |
            Required when pricing_model = revenue_share.
            Percentage of buyer's resale revenue paid to the producer. Range: 0–100.
        required_service_ids:
          type: array
          items:
            type: string
          description: |
            UUIDs of capabilities that buyers must have to submit an offer.
            Leave empty (or omit) for no restrictions.
      required: [material, quantity, unit, city]
...
    ListingRequiredService:
      type: object
      description: >
        A capability required by a listing producer. Buyers must have this capability
        (and optionally a license) to submit an offer.
      properties:
        id:
          type: string
        key:
          type: string
          description: Stable internal key matching capabilitiesTable.key.
        name_ar:
          type: string
        name_en:
          type: string
        requires_license:
          type: boolean
          description: If true, the buyer must also have license_status = approved.
      required: [id, key, name_ar, name_en, requires_license]
...