# lib/api-zod/src/generated/types/createWasteListingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tadweerah API
 * OpenAPI spec version: 0.1.0
 */
import type { PricingModel } from "./pricingModel";
import type { SaleType } from "./saleType";
import type { WasteMaterial } from "./wasteMaterial";
import type { WasteUnit } from "./wasteUnit";

export interface CreateWasteListingBody {
  material: WasteMaterial;
  /** @exclusiveMinimum 0 */
  quantity: number;
  unit: WasteUnit;
  /**
   * @minLength 2
   * @maxLength 80
   */
  city: string;
  /** @maxLength 500 */
  description?: string;
  /** @minimum 0 */
  price_hint?: number;
  /** Governs settlement mechanics. Immutable once published.
Defaults to "fixed" if not supplied.
 */
  pricing_model?: PricingModel;
  /** Controls offer model. Defaults to "auction" if not supplied.
   */
  sale_type?: SaleType;
  /** Optional FK to material_categories.id. */
  material_category_id?: string;
  /** Optional FK to material_categories.id (sub-level). */
  material_subcategory_id?: string;
  /** Optional FK to unit_options.id. Falls back to unit enum if absent. */
  unit_option_id?: string;
  /**
   * Required when pricing_model = revenue_share.
Percentage of buyer's resale revenue paid to the producer. Range: 0–100.

   * @minimum 0
   * @maximum 100
   */
  revenue_share_pct?: number;
  /** UUIDs of capabilities that buyers must have to submit an offer.
Leave empty (or omit) for no restrictions.
 */
  required_service_ids?: string[];
}