# lib/api-zod/src/generated/types/listingRequiredService.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tadweerah API
 * OpenAPI spec version: 0.1.0
 */

/**
 * A capability required by a listing producer. Buyers must have this capability (and optionally a license) to submit an offer.

 */
export interface ListingRequiredService {
  id: string;
  /** Stable internal key matching capabilitiesTable.key. */
  key: string;
  name_ar: string;
  name_en: string;
  /** If true, the buyer must also have license_status = approved. */
  requires_license: boolean;
}